move the helpers to a sal and make the rest of requests to redis aswell
